create test/tmp and ignore it, per ruby-build
since setup() is global, and most tests already have a local variant,
the global setup was being overridden. This commit (and previous revert
commit) bring us back in line with ruby-build. It stinks to use git for
keeping the test/tmp directory around. But it's the best option for now.