fix bad tar.gz test fixture
The binary.tar.gz tarball was malformed (missing gzip compression).
Somehow, it continued to extract correctly on OSX, but failed on travis'
linux. This error was previously being masked by the lack of error
bubbling on the pushd command! Yet another boon for the removal of
automatic compilation fallback which necessitated the manual error
bubbling.