jxcore dir structure needs fixed after each package install
Instead of running once after both jx and npmjx are installed, we need
to run it after *each* of them. This is because the steps in the npmjx
setup require a working jx exe. We don't have access to the jx exe in
the build dir while we're installin npm. And we can't use the installed
jx until it's moved there (by the 'fixe_jxcore_dir' function)

So this function will be invoked multiple times, but that's okay.