branch+pr per node version
This avoids the issue wherein the branch already exists on the remote.
If the branch name is consistent, then it will end up being re-used. And
since the branch can not be assumed to be culled or rebased, then a
fast-forward push is not guaranteed. And since we also don't want to
force-push to the branch, we need a non-static branch name. Rather than
using a timestamp or something, we may as well create a branch per
new-node-def and use the new node version as the branch name.