Allow versioning without publishing to brew
Defaults to publishing. To *not* publish, pass `--node-build:publish=`
with a value of anything other than 'true'.

The not-equal check is necessary (as opposed to checking for =true)
because we want full script to return a successful status code if we
aren't publishing (assuming the push was successful).

This way either the test is successful (thus the postversion script
returns 0) *or* the postversion script status is equal to the status of
the publish command. (Assuming in both cases that the push was already
successful.)