NODE_BUILD_MIRROR_CMD generate's mirror URL
Not all mirrors will use the identical scheme of matching the nodejs.org
url structure with `/dist` omitted. Now users may specify
`NODE_BUILD_MIRROR_CMD` to be the name of a command or function which
constructs the full mirrored package URL when given the original package
URL and checksum as arguments.

If omitted, the default scheme is used wherein NODE_BUILD_MIRROR_URL is
substituted for `https://nodejs.org/dist/`

The presence of *either* NODE_BUILD_MIRROR_URL or NODE_BUILD_MIRROR_CMD
is sufficient to trigger a mirror attempt.