Remove unnecessary uname stubbings
It was originally assumed that the uname stubbings were necessary and
the cause of the failures when rbenv/ruby-build#1038 was merged.

This is because `unstub uname` is what triggers the test failure.

However, all the reasons that make these stubbings necessary only apply
to ruby-build (we don't bother linking with readline).
Thus the uname check for FreeBSD that is necessary in ruby-build isn't
done here in node-build. Thus the uname stubbings fail, because uname is
never invoked.

So this change removes the uname stubbings that were introduced by
ruby-build, which caused the tests to break on macOS, whose existence is
only necessary because of the PATH change for ruby-build, but only
manifests on macOS.