Remove extraneous option fallback
The optarg for `-o` is required, so OPTARG won't be unset.
And if it's empty, we already have it defaulting to origin a few lines
further down so no need to do it twice. (Indeed, defaulting in two
places is confusing and could drift.)
The secondary default is required regardless, because the -o option
might have been omitted.