install hook to attempt retry without prefix
Node versions have a couple very common forms; with v, node-, or node-v
prefix. This install hook will attempt to retry a failed installation
_without_ these common prefixes.

It only attempts a retry if:
- the prior installation failed
- and it failed specifically due to not finding the definition file
- and the version name has one of 'v', 'node-' or 'node-v' prefixes

It maintains the flags passed to the original install command; (however,
it only maintains _known_ flags. That is, if nodenv-install learns some
new flags, this hook will need to be made aware of them)

This hook _execs_ the retry; which means this hook pre-empts any
subsequent hooks from the initial install. Presumably, this should be
safe since _most_ install hooks only run on success. Further, the retry
command replaces the currently-running process, which means the second
attempt's exit status will be the final status code of the original
invocation.