Remove npm caching
The npm cache is frequently over the 400mb limit that GitHub imposes,
which means the cache is rarely even saved.

And when it _is_ used, the caching only saves fractions of a second.
(probably negated by the cache action itself!!!)

Better to just simplify the config and remove the caching logic
entirely. I'm sure GitHub has npm registry proxying so the installs are
effecient anyway.