Distinguish failure modes
Not having changes to release or already being on the latest release are
"non-failure" exit modes, but must terminate as an error.

Being on the wrong branch or having unexpected changes (in strict mode)
are failure modes that are more "serious".

The intention is that the former (exit 1) will stop the GitHub workflow,
but report the job as a success. The latter (exit 2) will stop the
workflow and report the job as a failure.