Remove unnecessary stubbings
Most of the brew and uname checks in ruby-build are for
checking/compiling yaml, readline, gmp and openssl.

The only real uname/brew calls we have are for resolving the number of
cores, which is skipped when MAKE_OPTS are provided explicitly (as they
are in most of the stubbed make calls).